+++ packages/pangea-sdk/tests/integration2/test_ai_guard.py
@@ -8,7 +8,7 @@ import pytest
 from pangea import PangeaConfig
 from pangea.asyncio.services.ai_guard import AIGuardAsync
 from pangea.services import AIGuard
-from pangea.services.ai_guard import LogFields, Message, TextGuardResult, get_relevant_content
+from pangea.services.ai_guard import ExtraInfo, GuardResult, LogFields, Message, TextGuardResult, get_relevant_content
 
 from ..utils import assert_matches_type
 
@@ -42,6 +42,18 @@ class TestAIGuard:
         assert response.status == "Success"
         assert response.result
         assert_matches_type(TextGuardResult, response.result, path=["response"])
+
+    def test_guard(self, client: AIGuard) -> None:
+        response = client.guard(
+            input={"messages": [{"role": "user", "content": "hello world"}]},
+            recipe="foobar",
+            debug=True,
+            app_id="foobar",
+            extra_info=ExtraInfo(app_name="my app", foo="bar", baz="123"),
+        )
+        assert response.status == "Success"
+        assert response.result
+        assert_matches_type(GuardResult, response.result, path=["response"])
 
 
 class TestAIGuardAsync: