+++ packages/pangea-sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- AI Guard: multimodal guard.
+
 ## 6.6.0 - 2025-09-02
 
 ### Added